rename: 폴더 빈칸 제거 및 이름 소문자로 변환
