+++ src/pages/main/components/FilterList/FilterList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import CommonFilter from '../commonFilter/CommonFilter';
 import { ClassificationData, StudyCountData, technologyData } from './data';
-import TechnologyFilter from '../ technologyFilter/ TechnologyFilter';
+import TechnologyFilter from '../technologyFilter/TechnologyFilter';
 
 function FilterList() {
   return (
